backend/requirements.txt
# FastAPI e dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
motor==3.3.2  # MongoDB async driver
redis==5.0.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Processamento de dados
pandas==2.1.4
numpy==1.25.2
opencv-python==4.8.1.78
Pillow==10.1.0

# Machine Learning e MLOps
mlflow==2.8.1
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2

# Streaming e mensageria
kafka-python==2.0.2
celery==5.3.4
flower==2.0.1

# Monitoramento e observabilidade
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Utilitários
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
